bhmm output models c++ified

+++ sktime/markovprocess/bhmm/output_models/gaussian.py
@@ -19,7 +19,8 @@
 import numpy as np
 
 from sktime.markovprocess.bhmm.output_models.outputmodel import OutputModel
-from .impl_c import gaussian
+from ._bhmm_output_models import gaussian as gaussian
+
 
 
 class GaussianOutputModel(OutputModel):
@@ -118,7 +119,7 @@ class GaussianOutputModel(OutputModel):
         >>> p_o = output_model.p_obs(o_t)
 
         """
-        res = gaussian.p_obs(obs, self.means, self.sigmas, out=out, dtype=obs.dtype)
+        res = gaussian.p_obs(obs, self.means, self.sigmas, out=out)
         return self._handle_outliers(res)
 
     def fit(self, observations, weights):
